Change RuntimeError to SGMLParseError, which subclasses RuntimeError for backward compatibility.
Add support for SGML declaration syntax (<!....>) to some reasonable
degree.  This does not support everything allowed in SGML, but should
work with "real" HTML (internal subset in a DOCTYPE is not handled).
The content of the declaration is passed to the .handle_decl() method,
which can be overridden by subclasses.

--HG--
branch : legacy-trunk