Bumping to version 4.0 since we now support only XEmacs 21.1 and Emacs 20.7, although not all of the compatibility code for older Emacsen has been removed.  Specifically, the old "make sure we have a current custom.el library" stuff is removed, as is the hack-around for an NTEmacs 19.34.6 make-temp-name bug.
Updated much of the Commentary section in the initial comments.

Much more importantly, I've integrated Ken Manheimer's pdbtrack stuff,
which is way cool.  When enabled (as by default), this turns on the
overlay arrow when pdb is entered, either in the shell buffer or in
the *Python* buffer.  Specifically:

(py-mode-map): Added C-c C-d to toggle pdb tracking.

(py-pdbtrack-do-tracking-p): New user customizable variable to control
whether overlay arrow tracking is enabled or not.  This variable is
buffer local and is turned on by default.

(py-pdbtrack-minor-mode-string): The string that's added to the minor
mode alist when actually doing pdb overlay arrow tracking.  User
customizable.

(py-pdbtrack-toggle-stack-tracking, turn-on-pdbtrack,
turn-off-pdbtrack): New commands to control pdb tracking.

(py-pdbtrack-is-tracking-p): Helper variable used to control the
display of py-pdbtrack-minor-mode-string.  Set to true when the
overlay arrow is enabled, and false when it's disabled.

(py-pdbtrack-stack-entry-regexp, py-pdbtrack-input-prompt,
py-pdbtrack-track-range): Inherited from pdbtrack.el and renamed.

(py-pdbtrack-overlay-arrow, py-pdbtrack-track-stack-file): New
functions which actually do the tracking.

(py-shell): Add py-pdbtrack-track-stack-file to
comint-output-filter-functions.

Finally, add py-pdbtrack-track-stack-file to
comint-output-filter-functions at the file level.  This and the
py-shell addition should ensure that pdb tracking is installed
regardless of the order of operation.

Also, add py-pdbtrack-minor-mode-string to minor-mode-alist.

--HG--
branch : legacy-trunk